web/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
SQLAlchemy-serializer==1.4.1
Flask-Login==0.6.2
SQLAlchemy==1.4.18
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.4.4
Flask-SocketIO==4.3.2
python-engineio==3.13.2
python-socketio==4.6.0
Authlib==1.2.0
requests==2.28.2
Pillow==9.4.0